requirements.txt
numpy==1.26.4
opencv-python
imutils
dlib
pillow
PyYAML
deepface
onnxruntime
tqdm
